refactor(dynamic-forms)!: collapse PropertyDerivationOrchestrator into DerivationOrchestrator
The two orchestrator classes had the same lifecycle (constructor injects services,
wires onChange + debounced + HTTP + async streams) and after the previous commits'
deduplication only differed in the type-specific bits: which collection they
derive, what they pass to apply, and which sink they write into. Two classes for
the same shape was carrying its weight in mental overhead and DI provider sprawl,
not in code shared.

This commit makes `DerivationOrchestrator` the single class. Both pipelines opt
in via the constructor config:

  * `form` + `derivationLogger` set → value pipeline active (writes into
    FieldTree)
  * `propertyStore` set → property pipeline active (writes into store)

A single instance can run either or both. The DI provider in
`dynamic-form-di.ts` now constructs one instance with both pipelines wired —
previously each form bootstrap created two separate orchestrator instances with
nearly-identical RxJS state.

BREAKING CHANGES (under chore/1.0-cleanup so behind the 1.0 boundary):

  * `PropertyDerivationOrchestrator` class is removed. Callers used to invoke
    `new PropertyDerivationOrchestrator({ schemaFields, formValue, store })`;
    use `new DerivationOrchestrator({ schemaFields, formValue, propertyStore })`
    instead. The config field rename is the only API-shape change.
  * `createPropertyDerivationOrchestrator()` factory removed. Use
    `createDerivationOrchestrator()`.
  * `PropertyDerivationOrchestratorConfig` interface removed. Use
    `DerivationOrchestratorConfig` (with `propertyStore` instead of `store`).
  * `PROPERTY_DERIVATION_ORCHESTRATOR` DI token preserved as a `useExisting`
    alias for `DERIVATION_ORCHESTRATOR`. Existing `inject(PROPERTY_DERIVATION_ORCHESTRATOR)`
    calls keep working — they resolve to the same unified instance.

The legacy spec (`property-derivation-orchestrator.spec.ts`) is preserved and
exercises the property pipeline on the unified class — confirming that none of
its 30+ test cases regressed.

Numbers:
  * dynamic-forms fesm2022 bundle: 1,059,209 → 1,059,467 bytes (+258, +0.02%)
  * Source LOC: -62 net (~448 deleted, ~386 added)
  * 3062 / 3062 unit tests pass

Bundle is essentially flat. The actual bundle win — making the property pipeline
tree-shakable for apps that don't use it — requires lifting the property
providers into a separate `withPropertyDerivations()` feature factory (so apps
that don't import it skip the property collector/applicator/streams entirely).
That's the natural follow-up to this PR.

Why ship it anyway:
  * One class to grok, one instance per form, one DI provider entry to maintain.
  * Removes the conceptual fork between two pipelines that always run side-by-side.
  * Sets up the tree-shake follow-up cleanly (the property block is now a single
    `if (propertyPipelineEnabled)` branch inside one constructor).
